requirements.txt
torch>=2.2.0
numpy>=1.26.0
pandas>=2.2.0
python-chess>=1.999
kagglehub>=0.2.9
scikit-learn>=1.4.0
pyarrow>=14.0.0
tqdm>=4.66.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
optuna>=3.6.0
pysnooper>=1.2.0
gymnasium>=0.29.0
